+++ src/libslic3r/SLA/Hollowing.hpp
@@ -45,7 +45,7 @@ struct DrainHole
     bool is_inside(const Vec3f& pt) const;
 
     bool get_intersections(const Vec3f& s, const Vec3f& dir,
-                           std::array<std::pair<float, Vec3d>, 2>& out) const;
+                           std::array<std::pair<float, Vec3f>, 2>& out) const;
     
     Contour3D to_mesh() const;
